requirements.txt
smbus2
RPi.GPIO
gpiozero
cachetools